recruitMapper Test

+++ src/main/java/recruit/service/CreateRecruitService.java
@@ -58,8 +58,8 @@ public class CreateRecruitService implements RecruitService{
 		recruit.setTitle((String)map.get("title"));
 		recruit.setDescription((String)map.get("description"));
 		recruit.setCategoryId((int)map.get("categoryId"));
-		recruit.setDateStart((long)map.get("dateStart"));
-		recruit.setDateEnd((long)map.get("dateEnd"));
+		recruit.setDateStart((int)map.get("dateStart"));
+		recruit.setDateEnd((int)map.get("dateEnd"));
 		recruit.setHost((String)map.get("host"));
 		recruit.setUserId((int)map.get("userId"));
 		recruit.setMaxPersonnel((int)map.get("maxPersonnel"));
